Moved text helper to different directory
