+++ test/Intent/CancelIntentTest.php
@@ -17,7 +17,7 @@ use TravelloAlexaLibrary\Intent\CancelIntent;
 use TravelloAlexaLibrary\Intent\IntentInterface;
 use TravelloAlexaLibrary\Request\RequestType\RequestTypeFactory;
 use TravelloAlexaLibrary\Response\AlexaResponse;
-use TravelloAlexaLibraryTest\Application\TestAsset\Helper\TestTextHelper;
+use TravelloAlexaLibraryTest\TextHelper\TestAsset\TestTextHelper;
 
 /**
  * Class CancelIntentTest
