+++ test/TextHelper/TestAsset/TestTextHelper.php
@@ -9,14 +9,24 @@
  *
  */
 
-namespace TravelloAlexaLibraryTest\Application\TestAsset\Helper;
+/**
+ * PHP Library for Amazon Alexa Skills
+ *
+ * @author     Ralf Eggert <ralf@travello.audio>
+ * @license    http://opensource.org/licenses/MIT The MIT License (MIT)
+ * @link       https://github.com/travello-gmbh/amazon-alexa-skill-library
+ * @link       https://www.travello.audio/
+ *
+ */
 
-use TravelloAlexaLibrary\Application\Helper\AbstractTextHelper;
+namespace TravelloAlexaLibraryTest\TextHelper\TestAsset;
+
+use TravelloAlexaLibrary\TextHelper\AbstractTextHelper;
 
 /**
  * Class TestTextHelper
  *
- * @package TravelloAlexaLibraryTest\Application\TestAsset\Helper
+ * @package TravelloAlexaLibraryTest\TextHelper\TestAsset
  */
 class TestTextHelper extends AbstractTextHelper
 {
